Refactor: Volunteering controllers
A new base controller has been added containing the
current_volunttering_role_id method. The destroy
controller & the role controller now inherit from it.

The previous base controller has been renamed to the role
controller to better reflect its actual purpose.

The experience controller has been renamed to the start controller
as this is the first endpoint hit at the start of the unpaid
experience flow
