+++ app/controllers/candidate_interface/volunteering/destroy_controller.rb
@@ -1,7 +1,5 @@
 module CandidateInterface
-  class Volunteering::DestroyController < CandidateInterfaceController
-    before_action :redirect_to_dashboard_if_submitted
-
+  class Volunteering::DestroyController < Volunteering::BaseController
     def confirm_destroy
       current_experience = current_application.application_volunteering_experiences.find(current_volunteering_role_id)
       @volunteering_role = VolunteeringRoleForm.build_from_experience(current_experience)
@@ -20,12 +18,6 @@ module CandidateInterface
         current_application.update!(volunteering_completed: false)
         redirect_to candidate_interface_review_volunteering_path
       end
-    end
-
-  private
-
-    def current_volunteering_role_id
-      params.permit(:id)[:id]
     end
   end
 end
