+++ config/routes.rb
@@ -278,14 +278,14 @@ Rails.application.routes.draw do
       end
 
       scope '/unpaid-experience' do
-        get '/' => 'volunteering/experience#show', as: :volunteering_experience
-        post '/' => 'volunteering/experience#submit'
+        get '/' => 'volunteering/start#show', as: :volunteering_experience
+        post '/' => 'volunteering/start#submit'
 
-        get '/new' => 'volunteering/base#new', as: :new_volunteering_role
-        post '/new' => 'volunteering/base#create'
+        get '/new' => 'volunteering/role#new', as: :new_volunteering_role
+        post '/new' => 'volunteering/role#create'
 
-        get '/edit/:id' => 'volunteering/base#edit', as: :edit_volunteering_role
-        patch '/edit/:id' => 'volunteering/base#update'
+        get '/edit/:id' => 'volunteering/role#edit', as: :edit_volunteering_role
+        patch '/edit/:id' => 'volunteering/role#update'
 
         get '/review' => 'volunteering/review#show', as: :review_volunteering
         patch '/review' => 'volunteering/review#complete', as: :complete_volunteering
